+++ examples/weather/stormcast/utils/trainer.py
@@ -33,6 +33,7 @@ from utils.nn import (
     regression_loss_fn,
     get_preconditioned_architecture,
     build_network_condition_and_target,
+    unpack_batch,
 )
 from utils.plots import validation_plot
 from datasets import dataset_classes
@@ -101,6 +102,7 @@ def training_loop(cfg):
 
     background_channels = dataset_train.background_channels()
     state_channels = dataset_train.state_channels()
+    lead_time_steps = dataset_train.lead_time_steps
 
     sampler = InfiniteSampler(
         dataset=dataset_train,
@@ -140,7 +142,7 @@ def training_loop(cfg):
         regression_net = Module.from_checkpoint(cfg.model.regression_weights)
         if cfg.training.compile_model:
             regression_net = torch.compile(regression_net)
-        regression_net = regression_net.to(device)
+        regression_net = regression_net.eval().requires_grad_(False).to(device)
     else:
         regression_net = None
 
@@ -174,6 +176,9 @@ def training_loop(cfg):
         conditional_channels=num_condition_channels,
         spatial_embedding=cfg.model.spatial_pos_embed,
         attn_resolutions=list(cfg.model.attn_resolutions),
+        lead_time_steps=lead_time_steps,
+        amp_mode=enable_amp,
+        **cfg.model.get("hyperparameters", {}),
     )
 
     net.train().requires_grad_(True).to(device)
@@ -236,14 +241,14 @@ def training_loop(cfg):
         for _ in range(num_accumulation_rounds):
             # Format input batch
             batch = next(dataset_iterator)
-            background = batch["background"].to(device=device, dtype=torch.float32)
-            state = [s.to(device=device, dtype=torch.float32) for s in batch["state"]]
+            (background, state, lead_time_label) = unpack_batch(batch, device)
 
             with torch.autocast("cuda", dtype=amp_dtype, enabled=enable_amp):
                 (condition, target, reg_out) = build_network_condition_and_target(
                     background,
                     state,
                     invariant_tensor,
+                    lead_time_label=lead_time_label,
                     regression_net=regression_net,
                     condition_list=condition_list,
                     regression_condition_list=cfg.model.regression_conditions,
@@ -253,6 +258,7 @@ def training_loop(cfg):
                     images=target,
                     condition=condition,
                     augment_pipe=augment_pipe,
+                    lead_time_label=lead_time_label,
                 )
 
             if log_to_wandb:
@@ -303,15 +309,14 @@ def training_loop(cfg):
             batch = next(valid_dataset_iterator)
 
             with torch.no_grad():
-                background = batch["background"].to(device=device, dtype=torch.float32)
-                state = [
-                    s.to(device=device, dtype=torch.float32) for s in batch["state"]
-                ]
+                (background, state, lead_time_label) = unpack_batch(batch, device)
+
                 with torch.autocast("cuda", dtype=amp_dtype, enabled=enable_amp):
                     (condition, target, reg_out) = build_network_condition_and_target(
                         background,
                         state,
                         invariant_tensor,
+                        lead_time_label=lead_time_label,
                         regression_net=regression_net,
                         condition_list=condition_list,
                         regression_condition_list=cfg.model.regression_conditions,
@@ -328,6 +333,7 @@ def training_loop(cfg):
                         images=target,
                         condition=condition,
                         augment_pipe=augment_pipe,
+                        lead_time_label=lead_time_label,
                         **loss_kwargs,
                     )
 
@@ -337,6 +343,7 @@ def training_loop(cfg):
                             condition,
                             state[1].shape,
                             sampler_args=dict(cfg.sampler.args),
+                            lead_time_label=lead_time_label,
                         )
                         if "regression" in condition_list:
                             output_images += reg_out
@@ -416,9 +423,7 @@ def training_loop(cfg):
             fields += [f"steps {total_steps:<5d}"]
             fields += [f"samples {total_steps * batch_size}"]
             fields += [f"tot_time {current_time - start_time: .2f}"]
-            fields += [
-                f"step_time {(current_time - train_start - valid_time) / train_steps: .2f}"
-            ]
+            fields += [f"step_time {(current_time - train_start) / train_steps: .2f}"]
             fields += [f"valid_time {valid_time: .2f}"]
             fields += [
                 f"cpumem {psutil.Process(os.getpid()).memory_info().rss / 2**30:<6.2f}"